core: AddVmFromSnapshot - avoid early end method run
When adding vm as a clone of snapshot, the vm id should be set as
the entity id for the child commands parameters - this prevent a
situation in which the end method is called too early.

NOTE: seems like the MoveOrCopyImageGroupParameters class might be
re-factored so that in a case of given ContainerId, it will be set as the
entity id - but that's not in the scope of this change.

Change-Id: I7a536cb3c530fd9caa0b121830481604c0558d53
Signed-off-by: Liron Aravot <laravot@redhat.com>