core : ImportVm with snapshots without collapse - not all disks are copied
When importing a vmlists of disk ids and image ids are being populated
in order to be used when creating the parameters for MoveOrCopyImageGroupCommand.
When vm has snapshots, vm.getImages() returnes the DiskImage instances
of all the disks snapshots as well - that caused that when importing the vm
that has snapshots without collapse, same disk id was added few times to the disk's
id list.

When creating the tasks for copying the disks, the iteration is
according to the disk count, meaning that possibly we will execute copy
of the same image group while other image groups won't be copied at all
- which will cause that the imported vm will have only some of it's disks.

the proposed fix adds only the disk id\image of the leaf disk, so
duplicate id's won't be added.

Change-Id: Ib850d2d22ea6acb37320ad2b1bd6c571e1048faa
Bug-Url: https://bugzilla.redhat.com/912697
Signed-off-by: Liron Aravot <laravot@redhat.com>