tools: Use JBoss modules instead of class path
Currently in order to run the tools we build the class path using mostly
the "build-classpath" script and also some ad hoc searches. This tends
to fail in environments where the jar files are installed in locations
different than expected. This patch changes the scripts that starts
the tools so that they use the JBoss modules launcher instead, thus
manual class path building is no longer required.

Change-Id: Id3e4f7a4bd9a3b31ceb453f9257b2ad8eaf03f5a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>