webadmin,userportal: fix Events tabs sorting
As a result of commit 2c5d00aac4f588260bb36f7a28d7be5e856a8474
the webadmin gets stuck (must be refreshed) when trying to
navigate back from any events tab/sub-tab.

The problem is that the method setItems is updated with
null when switching tabs - causing a NPE in the sort method.
The method setItems is actually redundant in EventListModel
and should be added instead to UserPortalVmEventListModel
and UserPortalTemplateEventListModel.

EventListModel -> SyncSearch calls setItems with an empty list.
It uses the refreshModel method for fetching events and
UpdateItems for reverse sorting them; i.e. the sorting in setItems
is effective only for classes that override the refreshModel
method. Therefore, moved the sorting to these classes and
added the nullity check.

[ToDo: the reverse sorting in EventListModel -> UpdateItems
should probably be re-factored in a way that allows sorting
for all inherited classes without overrides...]

Change-Id: Ibefebeea4dd804f6f930ed61e5899fe460996a70
RelatedTo-Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=907421
Signed-off-by: Daniel Erez <derez@redhat.com>