userportal: fix possible NPE in % progress bar
DoublePercentageProgressBar unboxes valueA and valueB (into fakeA and
fakeB, respectively), and only then checks if they were null.

If either of them is indeed null, we will get a NullPointerException
from the outboxing, and won't even reach the null check.

This patch moves the relevant code into the null-safe condition.

Change-Id: Id7a65802cc79a698863899085189ea80fcc6d57f
Signed-off-by: Allon Mureinik <amureini@redhat.com>