ui: DiskModelByAliasComparer: null disk treatment
DiskModelByAliasComparer was checking DiskModel.getDisk() for being
null, but then ignoring the check, and performing the same operation
(DiskModel.getDisk().getDiskAlias()) regardless.

This is obviously wrong, since if the disk is null, this will result in
a NullPointerException. This patch assigns a default (empty String) as
the disk alias in case the disk is null.

Change-Id: I2d060a5b9d87dc3ec7bcc382d90a4f2825ca9a05
Signed-off-by: Allon Mureinik <amureini@redhat.com>