packaging: cleaning up dal's pom.xml
Moved the test resource from 'resources' to 'testResources' section
and removed the resource section.

And this will also make sure log4j.xml not gets into dal.jar.
If the log4j.xml included in the dal.jar, after the installation
the log output is not going to engine.log and it always remains
empty.

Change-Id: Ibce1ba5034ef9d9043a7919c53d711f087105dc2
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>