packaging: Creating keystore for jboss use
Currently jboss uses apache.p12 keystore file. In cases
where apache is not used for proxying, jboss tries to
use this file directly and fails on permissions.

This patch tries to resolve the issue by adding an additional
keystore jboss.p12 during setup/upgrade for the use by the jboss.

Change-Id: I22d71d9de011e8af4bde26d9e2a048a6387ce70f
Signed-off-by: Alex Lourie <alourie@redhat.com>