packaging: Updated taskcleaner use in remote DB flows
Currently taskcleaner is executed as if it was running only
on a locally installed DB. Added the correct parameters with
values of host/port of the DB, so that in case of remote DB
all the values will be sent correctly and the script will
execute on a remote DB server.

Additional items required to handle remote DB connections:

* Added missing common.sh file to the taskcleaner
packaging.
* Updated deployment of the async tasks sql for the remote DB
* Updated dbfunctions to use ${USERNAME} instead of 'postgres'
user to allow remote connections work.

Change-Id: Ie3c205962521bafe56ff67f3f46dd83202f48b7e
Bug-Url: https://bugzilla.redhat.com/915028
Signed-off-by: Alex Lourie <alourie@redhat.com>