restapi: Use context classloader from YAML
The YAML library that we use to build load the RSDL metadata assumes
that it can load classes using from its own class loader, and this is
incorrect, specially now the the library is loaded from a JBoss module.
The correct way to load classes is to use the context class loader, so
it will work regardless of where in the application the classes are
located.

Change-Id: Ie86d591a0ce71b9984fa9c76ee42de67dce2ae4e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>