restapi: Use context classloader from YAML
The YAML library that we use to build load the RSDL metadata assumes
that it can load classes using from its own class loader, and this is
incorrect, specially now the the library is loaded from a JBoss module.
The correct way to load classes is to use the context class loader, so
it will work regardless of where in the application the classes are
located.

Change-Id: Iad370caa7940c3c5883bc21b34d1eea73cfc63d5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 2fd239a065877c20ea580926e28d6603f6fe1d75)