core: Fix quota check for Add-Template
It is possible to create a template even if quota has no sufficient
resources for the disk. The internal disk list in the command was
created only after the quota consumption parameters were checked. This
only happens when calling the command from REST API since the UI
initialize the disk list.

The getQuotaStorageConsumptionParameters() method was changed so that if
the internal disk list was not initialized it will be taken directly
from the vm disk list.

Change-Id: I48bffa41aa8a049a184b8cde262ad9c7f9b3baff
Bug-Url: https://bugzilla.redhat.com/913551
Signed-off-by: Ofri Masad <omasad@redhat.com>