core: add the image block alignment scan
This patch introduces the new command required to scan the block
alignment of disks (both images and direct LUNs).
The new XML-RPC call issued to VDSM is:

 scanDiskAlignment(vmId, driveSpecs)

vmId is the UUID of the VM that is attached to the disk or a blank
UUID in case of a floating disk (not included in this patch yet).
In the future the vmId can be used to identify a running VM and
execute the scan through a guest agent in those cases where it's
impossible to do it concurrently (e.g. COW disk format).
The driveSpecs parameter is in the same format used when starting
a VM on a VDSM host:

 driveSpecs = {
   'poolID':   <poolID>,
   'domainID': <domainID>,
   'imageID':  <imageID>,
   'volumeID': <volumeID>,
 }

Or for direct LUNs:

 driveSpecs = {
   'GUID': <GUID>,
 }

The expected information returned in case of success is:

 alignmentInfo = {
   'status': {'message': 'Done', 'code': 0},
   'alignment': {
      '/dev/sda1': True,
      '/dev/sda2': True,
 }}

The result (with the scan execution timestamp) is stored in the
base_disks table (as other similar properties such as the bootable
and shareable flags).

The feature page can be found at:

 http://www.ovirt.org/Features/DiskAlignment

Change-Id: I4858b7bbfa453230fcafecfbc5358c715d5d825b
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>