restapi: trace export template command by job status
This patch prevents a situation in which the REST reports that export
template command was finished before the end-action event is added.

This situation occured because the REST part monitored the command
according to the statuses of its VDSM tasks - when they were done,
it reports back to the user that the command has finished and gave
him back the control (the command was invoked synchronously),
sometimes before the engine managed to add the end-action event.

This patch change the way the export template command is monitored by
the REST to be according to the command's job status, that way we ensure
that the end-action event is added before REST is reporting back to the
user that the command is finished.

Change-Id: I4442b337e0242aa756825243dbcf6db1a2b0559f
Bug-Url: https://bugzilla.redhat.com/860194
Signed-off-by: Arik Hadas <ahadas@redhat.com>