packaging: Updated psql calls not to ask password
Currently we use our own pgpass file to supply the required
password for the user. This patch adds a "-w" flag for all
psql calls that inhibits quering password at all. This will
lead to psql connection errors if pgpass was incorrectly
defined, but password will still not be asked, which is the
desired behavior of the setup.

Bug-Url: https://bugzilla.redhat.com/915409
Change-Id: Ib73a7e3a5955a66afa0eeeafa80cf1f6464e8fd4
Signed-off-by: Alex Lourie <alourie@redhat.com>