packaging: Move dependencies to a JBoss module
Currently we package dependencies as .jar files in the lib directories
of the .ear and in the .war of RESTAPI. This has the inconvenient that
those dependencies don't have an associated artifact and also can't be
shared from different applications. This patch moves all those
dependencies to the dependencies jboss module, so that they are added to
the corresponding artifact, and can then be easily installed and shared.

Change-Id: I9a211b4c24dd02c6e018c6be62beee2d1f5f9afe
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit e8081ab1b6e560ada7a26240e97dfc65ec475423)