tools : Unable to add remote domain using engine-manage-domains
When there is an error with engine config while adding a remote
domain the toll does not log the error to the log file.

This patch sets the default log level of engine config to DEBUG
instead of OFF. This write the debug messages to the log file
instead of supressing it.

Change-Id: I8457b778274acabaab30b27b0605b8600e338873
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=912736
Signed-off-by: Ravi Nori <rnori@redhat.com>