core: Remove rmw-war
This war, located in the frontend, contains a few servlts that are in
fact more related to the backend, and it doesn't grant the existence of
a different module and .war file. This patch moves all the content to
the root web application. As the URLs don't change there is no impact
outside of the engine.

Change-Id: I1b5a292cb28cbc460cf45a9595bddd7ebb824857
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>