webadmin: fix IAE on run once dialog
This patch fixes IllegalArgumentException that was thrown when trying to
run-once a VM with no bootable device.

Before change http://gerrit.ovirt.org/#/c/11735 the boot sequence that
was generated when trying to run-once a VM was never empty - it always
contained at least the hard-disk (even for diskless VM).
After applying the change above, the returned boot sequence for a VM
with no bootable device is empty, and wrong handling of the empty
string that represents it results in IllegalArgumentException.

Change-Id: I42197615331c6b62f5978e1d4546298aee617db4
Bug-Url: https://bugzilla.redhat.com/885994
Signed-off-by: Arik Hadas <ahadas@redhat.com>