engine: Hot-unplug fails when unplug nic and changing MAC address.
Hot-unplug fails when trying to unplug and change MAC address for nic
under running VM.
When unplugging a nic of a running vm-
1. The db is updated.
2. Unplug action is sent to the host.

It means that the vdsm gets unplug with the new mac address (although the
vnic is defined on the host with the old one).
To solve this issue the old mac address should be sent to the vdsm.

Change-Id: I185e5572501ea86443a93f4fdf309e0e05520646
Bug-Url: https://bugzilla.redhat.com/904840
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>