engine: Adding vds group name field to GlusterVolumeEntity
A new view named 'gluster_volumes_view' is added by joining
'gluster_volumes' and 'vds_groups' tables to get the cluster name
of the volumes. gluster volume functions are modified to use
the 'gluster_volumes_view' instead of 'gluster_volumes'.

'vdsGroupName' field is added to the GlusterVolumeEntity and
it will shown in the 'Cluster' column of Volumes tab in the UI.

Change-Id: I6b98e7b46e312d0dc0b9b29b7cb8a0b0a7ecd3b7
Bug-Url: https://bugzilla.redhat.com/875800
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>