webadmin: Added SuggestBox bond name widget
When creating a new bond or editing an existing one through the newer
bond popup window, the widget for picking a bond name is now a
SuggestBox rather than a ValueListBox. This enables the user to input a
new bond name, as well as choose between suggested bond names that
already exist on the host and are free (as was customary historically).

Change-Id: I130586dc4787abecf81566a8d3b723d093f6fca2
Signed-off-by: Lior Vernia <lvernia@redhat.com>