webadmin: Added default created bond to comply with SuggestBox
Since when performing a bonding action, the user may now create a
newly-named bond, it doesn't make much sense to choose one of the
existing free bonds by default. The solution is to put a "bond" prefix
as default, and let the user either proceed to add an arbitrary number
or choose one of the suggested existing bonds.

Change-Id: I6b68f6b5f89089a7f78609b78df2e3e92827234d
Signed-off-by: Lior Vernia <lvernia@redhat.com>