userportal: ignore the NetworkUser role
Since 97e84c622a7cfa9f709eb4d814d42d90448edc91 there is a
NetworkUser role and this role is assigned to everyone.

Since the engineUser does not contain this role, UP
evaluates, that this specific user has more roles than the
engineUser so he is an admin and should see the extended tab.

Fixed by ignoring the NetworkUser in this algorithm
the same way as e.g. the QuotaConsumer is.

Change-Id: Iaf696b51159ba016d88124b789d5bbbab7b736fe
Bug-Url: https://bugzilla.redhat.com/916328
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>