restapi: Omit of prefer header doesn't turn off session based authentication #916582
This patch makes sure we take the http session in case it exists, even
if the prefer auth header isn't set.
That way, the last request without the prefer header will succeed, and
then we would log out. Also, the first request with the prefer header
will create a new session, as expected. Requests without the prefer
header will not create a session, and no cookie will be returned.

Change-Id: Ie61285212c4050bc6dc2c744b3d281648ea542ca
Bug-Url: https://bugzilla.redhat.com/916582
Signed-off-by: Oved Ourfali <oourfali@redhat.com>