core,restapi: Alias for registered disk
Change-Id I82de498fd9a8e25ed9e1dc5776f2fdf0c35b46da introduced a new
operation, registering an unregistered disk (i.e., a disk that exists on
the storage domain but not in the engine's database).

Since disk aliases are an engine concept, the registered disk did not
have an alias, which is not a legal state in the engine.

This patch allows passing an alias when registering a disk, and
auto-generates an alias ("RegisteredDisk_{YYYY-MM-DD_hh-mm-ss}") if
it is not passed.

Change-Id: I6ab17da94ff7ff16852efb3a374148dacf1135f8
Signed-off-by: Allon Mureinik <amureini@redhat.com>