core:Upgrade from 3.1 to 3.2 fails
Upgrade failed on 3.1.3 to 3.2 on both PG 8.4.x and 9.x

On 8.4.x :

Problem was that attempt to drop the uuid generation support functions
was done with ${USERNAME}. since postgres is the owner of those
functions the user engine assigned to ${USERNAME} has no permissions to
drop those functions. Solution was to run the script that drops those
functions with user postgres.

on 9.x :

The DROP EXTENSION statement tried to drop uuid-ossp.
In postgres the '-' functions as '?' in regular expressions so I had to
wrap the name with double quotes.

Change-Id: I70eff71a8367879eae53ad3b003a050e289bc73e
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=916728