engine: Management network should be viewable on UserPortal
When creating an empty data-center, a management network
is created automatically for it, without any permissions.

This prevents from the users to use the network without
granting permissions explicitly for that network.
In order to improve the user-experience, the default network
will be granted for 'everyone' user with 'NetworkUser' role.

Change-Id: I3bf7a3c16b61bf343fe0994050042573e06b3c33
Bug-Url: https://bugzilla.redhat.com/902697
Signed-off-by: Moti Asayag <masayag@redhat.com>