core: no quota refresh when vm stopped
Currently the vm statuses on which we are calculating quota consumtion
isn't consistent with db (managed in 2 places).
In this patch I've sync-ed all places, that use vm status to decide whether
to count the quota, to a single stored proc that is being used by server
and other stored procs.

Change-Id: Ie2bd42a8431325ddbe552363b299fb80dbd7f3c5
Bug-Url: https://bugzilla.redhat.com/908834
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>