engine: Fix brick validation during Add Bricks
While adding a new brick(s) to a volume,
server_id and brick_directory combination will
be validated against the entries in the gluster_volume_bricks table.

If there are any existing entries found for the server:brick_dir,
an error message will be thrown to the user.

Introduced a new procedure to get a brick entity based on server_id
and brick directory.

Also an error message will be thrown if there any
duplicate bricks found.

Change-Id: I83ec15a183d81a274388c313006208d439809548
Bug-Url: https://bugzilla.redhat.com/856109
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>