core: Fix quota not inherited from VM to Template
When Template was created from a Vm using the REST API the quota of the
disks was not inherited by the new template disks.

The command was looking for quota to assign to the new disk in the
DiskInfoDestinationMap. But as quota is not yet fully supported in REST,
this map is left empty.

Added an 'else' block which uses the old disk quota as a default.

Change-Id: I697082483c706178b8ea47306779dc2b63da5ed6
Bug-Url: https://bugzilla.redhat.com/913551
Signed-off-by: Ofri Masad <omasad@redhat.com>