packaging: Add support for ${...} in local config
This patch adds support for references to parameters in the local
configuration file. This is convenient in order to simplify the
packaging of other applications, like reports, as with this support they
don't have to modify the existing configuration files, but add new ones
in the /etc/ovirt-engine/engine.conf.d directory. For example, in order
to add the reports application there is no need to modify the existing
engine.conf file to add the following:

  ENGINE_APPS=engine.ear reports.war

Instead of that a file can be created in engine.conf.d with the
following content:

  ENGINE_APPS=${ENGINE_APPS} reports.war

Change-Id: Ib8ce501241ed1a937db24e1371e2d4227679ae74
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>