core: Snapshot should not be update after rollback.
Once createAllSnapshotCommand execution encountered a rollback,
The new snapshot is deleted as part of the compensation scenario.

That will cause the end method to get null value for the snapshotId and
will cause an sql exception when it will be updated.

The update phase should be only when the tasks failed but
also when command finished successfully and the snapshot was already
created in the DB.

Change-Id: Ia96cd0958dcdb34d590a160b422b3b2562149344
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Bug-Url: https://bugzilla.redhat.com/917719