core: RFE: Multi-Tier fencing
This patch fixes a bug in the multiple sequential PM agent treatment.
The secondary agent was used only if the primary agent was configured OK
and the fence command was sent successfully to it and it fails to perform
the command after the configured wait/retries.
This patch change this behaviour to use the secondary agent also if the
primary agent was not configured well or from some reason can not be
used although having the correct configuration (can not be accessed for
example due to networking issue)

Change-Id: I8ee542e7ddaec1bcdc777295b8222aceed798ed9
Signed-off-by: Eli Mesika <emesika@redhat.com>