core: change can-do checks for resume action
Currently resume is done as a special flow of run-vm command when the
previous state of the VM is paused. the can-do-action checks in that
case were the same as those that are made for regular run-vm action (for
VM which are down).

This patch changes the can-do-action checks in case of resume action
such that they will only check that the VM exists. there is no need to
do further checks as the VM was already created (thus was found to be
valid before) if its state is paused.

Change-Id: I772377045d5c6d694e00006c9b12e210c52ab353
Bug-Url: https://bugzilla.redhat.com/885994
Signed-off-by: Arik Hadas <ahadas@redhat.com>