packaging: Use Id=... to check systemd services
The services managed by systemd can have an unique identifier and then
several alternative names, or aliases. The NFS service, for example, has
"nfs-server" as identifier but it also has "nfs" and "nfs-server" as
aliases. These alternative names can be used for some operations, but
not for all of them, in particular the "systemctl enable ..." operation
requires the main name, it will not work with alternative names. This
meas that we need to make sure that the method that checks the
availability of services uses the main name, otherwise it can succeed
but then later the call to "systemctl enable" can fail.

Change-Id: I60ba0a5481f191c67eeb1fed6abd3f961cfbc8f4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit b2f6cbeac1cef232b560269faa50de4d030bb282)