packaging: Use Id=... to check systemd services
The services managed by systemd can have an unique identifier and then
several alternative names, or aliases. The NFS service, for example, has
"nfs-server" as identifier but it also has "nfs" and "nfs-server" as
aliases. These alternative names can be used for some operations, but
not for all of them, in particular the "systemctl enable ..." operation
requires the main name, it will not work with alternative names. This
meas that we need to make sure that the method that checks the
availability of services uses the main name, otherwise it can succeed
but then later the call to "systemctl enable" can fail.

Change-Id: I99492161d29254e48a0196b3b5cbdfa7c9dc1863
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>