core: extract null check from VmRunHandler#canRunVm
This patch extract the check whether the given VM is null from the
VmRunHandler#canRunVm method. the reason for that is that in both places
that call this method it makes sense to check it before making the call.

So now as part of the "contract" of this method, the given VM must not
be null (and the method's documentation was updated accordingly).

Change-Id: I42fd8274577ba94accf16a4f944de07ca9ea5419
Signed-off-by: Arik Hadas <ahadas@redhat.com>