core: Power Management or fence errors are not...
Power Management or fence errors are not logged in EVENT tab of GUI

Adding a call to AlertIfPowerManagementOperationSkipped method in
handleError method.
This will force raising an Alert each time a PM operation was supposed to
be executed and from some reason the fencing operation has been failed.

In the case of dual agents:
for concurrent - on Start , Alert will be generated if both agents
failed
                 on Stop , Alert will be generated if one agent failed
for sequential
                 on start/stop , Alert will be generated if both agents

Change-Id: If5aef618f1e1ee8ab282bad528218b15b61085bd
Signed-off-by: Eli Mesika <emesika@redhat.com>