core: disable everyone for login
If the "everyone" has assignd a role which contains login,
from this point everyone is allowed to log in.

In this specific bug the NetworkUser has been assigned to
everyone which contained the login. This caused that everyone
was allowed to log in. This issue did not happened the first time
so this patch solves it in a more generic way - ignores the
everyone when checking if the user is allowed to log in.

Change-Id: I7bba37a450efdd9e3475470e4ed8f49347c51a2e
Bug-Url: https://bugzilla.redhat.com/916328
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>