core: remove duplicate event on successful hibernation
Successful hibernation (suspension) of VM resulted in three events:
1. event that indicates that the action was initiated
2. event that indicates that the action was completed
3. event that indicates that the VM is suspended

The second event was generated by the command and the third event was
generated by the monitoring. those two messages are expressing the same
thing so we don't need both of them.

Moreover, the second event is misleading because the command generates
it right after sending the request to VDSM, thus it can't say that the
action is completed at that point. so this patch removes the second
event.

Change-Id: I5445374ddf6996e696c71995fe3bd2a9ddd60a85
Bug-Url: https://bugzilla.redhat.com/918306
Signed-off-by: Arik Hadas <ahadas@redhat.com>