core: UpdateVmDisk: unify other disk retrieval
Several validations in the canDoAction() rely on getting the VM's other
disks from the database, which is unfortunately done in the said several
places.

This patch unifies them all to a lazy getter, getOtherVmDisks() for
better performance and readability.
Besides the code change, it also introduces a couple of new unit tests
to ensure that no functionality was damaged.

Relates-To: https://bugzilla.redhat.com/854964
Change-Id: I31bcb767bf86d6f16116cfa06a710bb0488b087f
Signed-off-by: Allon Mureinik <amureini@redhat.com>