engine: Passing to ActivateDeactivateVnic the vnic
As ActivateDeactivateVmInterface became an internal command it shouldn't
get the vnic from the db.
The command that called the ActivateDeavtivate have already got the vnic
and can pass it.

In case the caller command is UpdateVmInterface and the action is unplug,
the passed nic should be the old nic (the host is familiar to)
and not the updated one.

Change-Id: I058b7197ec5a152e91cbb3c362b290326678df40
Bug-Url: https://bugzilla.redhat.com/904840
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>