engine: Added checks for external networks
1. A check that makes sure the imported network is new in the data
center, otherwise it makes no sense to import it twice since it's the
same underlying network.

2. A check that makes sure that only name or description can be changed
on an external network, since the other details don't reflect what the
properties of the network on the external provider.

3. A check that makes sure external networks are added as VM networks.
This is because we don't know what is an external non-VM network and how
to support it, so currently we block it.

Change-Id: I0a02f3caffa5b2d1d36e908f79db0c788fff5a8f
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>