core: Failed install DB during oVirt-setup...
Failed install DB during oVirt-setup not on clean server

This patch adds a utility encodingvalidator.sh that can be used to
validate that template1 database encoding is UTF8
If template1 database has any other encoding, giving this utility the -f
flag will fix the problem.

Since we were not able to reproduce the BZ, it was agreed that supplying
this utility and calling it from the installer in the
pre-install/upgrade step is sufficient to resolve this problem.

Change-Id: I74dd62ec2c0f96d719b15848da4e08d8f6cc5068
Signed-off-by: Eli Mesika <emesika@redhat.com>