engine: Prevent management network change
The patch prevents changing the management network address if
the host was added to the system with its IP address as the
computer-name for the certification creation.

The action is blocked in order to prevent connectivity loss from
the engine to the host. The host should be reinstalled in order to
modify the management network address.

Change-Id: If98a5853385ad484dfa6e5392797f96daeaaa381
Signed-off-by: Moti Asayag <masayag@redhat.com>