engine: Manage Gluster hooks support
This patch adds the database tables and stored procs
to manage gluster hooks.The associated Java entity
classes and DAO classes are added as part of this
patch.

For each cluster, the list of available hooks are
stored in the database. Under cluster, a hooks tab
will be provided to manage these.

The hooks table stores a status field which indicates
if the hooks are in sync across all servers. This is a
3 bit representation with each bit indicating the
conflict type - status conflict/content conflict/missing

Change-Id: If8b2d261ead720fccea6884acbc806ec79e52b36
Signed-off-by: Selvasundaram <sesubram@redhat.com>
Signed-off-by: Sahina Bose <sahina@gmail.com>