tools: Don't use console for debug log
Currently we use a console appender for the debug log level of the
engine-config tool, this means that the debug messages are mixed
with the regular output of the tool, and that confuses its users,
including engine-setup and engine-manage-domains.

This change introduces a new ToolConsole class for the tools. This
class centralizes all input and output to the console and sends all the
text to the log as well. The engine-config tool uses this new class for
regular input and output.

This change also removes the console appender from the log4j
configuration and changes the default log level to DEBUG so that debug
messages will be sent to the log as they used to be.

Change-Id: I556e20f6333e22faeb7d212767ebefb99edce54e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>