gluster: added cluster level check
refreshHeavyWeightData uses the vdsm verb glusterVolumeStatus
which is not available in cluster level 3.1. So added a check
in GlusterManager to not perform the refresh in case of a 3.1
cluster.

Also enhanced the AddVdsGroup and UpdateVdsGroup commands to
throw an error if their compatibility version doesn't support
gluster service.

Created a separate GlusterFeatureSupported class in it's own
package so that it can be separated out easily in future.

Change-Id: Id3944e8b1d8d82eb19643b99f606da2364d6b582
Bug-Url: https://bugzilla.redhat.com/920575
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>