gluster: Use brick server for advanced details
Currently, there is a bug in glusterfs, because of which the command
'gluster volume status' for a particular brick works properly only if
it is executed on the server where the brick resides.

While that bug gets fixed in glusterfs, I am changing the code here
to make sure that we always execute the command on the brick's server
(provided it is UP).

Another refactoring done in this patch is to expect volume and brick
IDs instead of names in GlusterVolumeAdvancedDetailsParameters

Change-Id: Id4acb0bbe84bacab714926d69b639bbda970df9a
Bug-Url: https://bugzilla.redhat.com/877961
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>