core: prevent VM creation from Template while removing its disks
In what seems to be a hack, we set a flag for not checking template
images in some commands. the problem is that the template images might
be locked for a good reason - like in case the images are really being
removed.

This patch prevents it from happening for addVmFromTemplateCommand: when
removing a disk we now take exclusive lock for that disk for the whole
execution of the command (RemoveDiskCommand). at AddVmFromTemplate we
now take shared lock for each of the template's disks - thus it won't
be possible to acquire the locks if one of the disks is being removed.

Change-Id: I8736751bbfea927072b2ade91ee3d9511e266501
Signed-off-by: Arik Hadas <ahadas@redhat.com>