core: canDoAction msg of locked disk being removed
Change the can-do-action message which is attached to locks that are
taken when removing disk in order to make it more informative than
the default generic message that is being returned when those locks
can't be acquired.

This change of the can-do-action message allows us not to pass the locks
from RemoveDiskCommand to RemoveImageCommand - as those locks can remain
until the end of the execution of RemoveDiskCommand, there's no hurry to
free them anymore.

Change-Id: Ib20a2be29d01c04497473d168e7abbe7d1bf2db8
Signed-off-by: Arik Hadas <ahadas@redhat.com>