userportal: don't pass ctrl+alt+delete for win>7
Don't pass the ctrl+alt+del for win 7 or newer client.

This has already been implemented, but on the level of
ConsolePopupPresenterWidget which is not enogh - it was
executed only after opening the consonle popup. But if
the user pressed the connect button, without opening the
the console popup, the ctrl+alt+delete was passed to the guest.

This patch moves this logic to SpiceConsoleModel.

Change-Id: I09c45cb30577e84830b6e82d79d2bb7ef7455b1b
Bug-Url: https://bugzilla.redhat.com/918650
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>