core: unify the GetAllImagesList queries
This patch unifies and consolidates the queries to retrieve the list of
images from a storage domain. The previous separate queries were:

 GetAllImagesList
 GetAllIsoImagesList
 GetAllIsoImagesListByStoragePoolId
 GetAllFloppyImagesList
 GetAllFloppyImagesListByStoragePoolId

which have been consolidated in:

 GetImagesList (renamed from GetAllImagesList)
 GetImagesListByStoragePoolId

The type of image to list (iso, vfd) is now a parameter that can allow
also additional types (e.g.: disk) and a generic "All" to retrieve the
entire list.

This is one of the first steps toward image storage domains that can
store any type of images (iso, vfd, disk, etc...).

Change-Id: I079e3f1c5bfbcfd1dc09ebce41a327c253eb53ac
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>