core: Cannot confirm Host has been rebooted when..
Cannot confirm Host has been rebooted when having a single host in the system.

This patch removes the SPM Host mark in the DC in case that Host is
non-responsive and is the only Host in the DC.

This enables to put the Host in Maintenance and remove it if needed.

Two scenarios tested
1) Making the Host responsive again => Host is SPM back
2) Adding another Host => New Host is SPM

This patch is not handling the case in which the problematic Host has
running VMs, only the scenario reported in the BZ was handled

Change-Id: I8264d91d6dc5f33e4684d3bef86624b6f424a64b
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=837539
Signed-off-by: Eli Mesika <emesika@redhat.com>