core: helper methods to create locking pair
It is becoming more and more common to use the new infrastructure that
allows us to attach customized error message to a lock, that will be
shown in case other locks cannot be acquired because of that lock.

This patch adds helper methods in LockMessagesMatchUtils class for
creating the locking pair which contains locking group and the
customized error message.

Change-Id: Id0c854cfc3b6cb1b7f7f4907452e116af5fbf6bf
Signed-off-by: Arik Hadas <ahadas@redhat.com>