packaging: Stopped using backup.sh script for DB backup
This patch replaces the usage of the backup.sh script
for backing up the DB during upgrade/cleanup operations
with the direct pg_dump command with same parameters.

The reason for this is that the script itself delivered
with a different package than the setup utilities, so its
functionality is not available to the setup; this leads
the setup to fail in various flows.

Change-Id: Ic2b6f5c36e3005b5fdeeb3a5558414fe353d5e6e
Bug-Url: https://bugzilla.redhat.com/920712
Signed-off-by: Alex Lourie <alourie@redhat.com>