core:remove storage pool is failing in Psql exception...
ovirt-engine: remove storage pool is failing in Psql exception: Failed
to run compensation on startup for Command
org.ovirt.engine.core.bll.storage.RemoveStoragePoolCommand, Com mand Id
: 076abb88-ead3-4a87-9d31-a534f41ad127, due to: DataIntegrityViolatio

Since there may be some compensation entries that are not related to
Zombie tasks left and this may cause the backend to abort when trying
to activate the compensation procedure on system startup.

This patch adds a -A flag to taskcleaner utility that enable to remove
all compensation data from the business_entity_snapshot table.

Installer should call taskcleaner with the -A flag.
instead of running taskcleaner with  -zRCJq flags, installer should
invoke taskcleaner with the -zRAq flags

Change-Id: Iea501a3da1b0fe72b4c4e95b751d51cd6efb7843
Bug-Url: https://bugzilla.redhat.com/921202
Signed-off-by: Eli Mesika <emesika@redhat.com>